+++ examples/polymorphic/expected/Operations/AllMembers/AllMembersErrorFreeResult.php
@@ -7,4 +7,9 @@ namespace Spawnia\Sailor\Polymorphic\Operations\AllMembers;
 class AllMembersErrorFreeResult extends \Spawnia\Sailor\ErrorFreeResult
 {
     public AllMembers $data;
+
+    public static function endpoint(): string
+    {
+        return 'polymorphic';
+    }
 }