+++ examples/polymorphic/expected/Operations/NodeMembers/NodeMembersErrorFreeResult.php
@@ -7,4 +7,9 @@ namespace Spawnia\Sailor\Polymorphic\Operations\NodeMembers;
 class NodeMembersErrorFreeResult extends \Spawnia\Sailor\ErrorFreeResult
 {
     public NodeMembers $data;
+
+    public static function endpoint(): string
+    {
+        return 'polymorphic';
+    }
 }