+++ examples/polymorphic/expected/Operations/UserOrPost/Node/User.php
@@ -40,4 +40,9 @@ class User extends \Spawnia\Sailor\ObjectLike
             'name' => new \Spawnia\Sailor\Convert\NullConverter(new \Spawnia\Sailor\Convert\StringConverter),
         ];
     }
+
+    public static function endpoint(): string
+    {
+        return 'polymorphic';
+    }
 }