+++ src/Error/Location.php
@@ -1,0 +1,31 @@
+<?php declare(strict_types=1);
+
+namespace Spawnia\Sailor\Error;
+
+use stdClass;
+
+/**
+ * Beginning point of the syntax element in the GraphQL document associated with the error.
+ */
+class Location
+{
+    public static function fromStdClass(stdClass $location): self
+    {
+        $instance = new static();
+
+        $instance->line = $location->line;
+        $instance->column = $location->column;
+
+        return $instance;
+    }
+
+    /**
+     * Line number starting from 1.
+     */
+    public int $line;
+
+    /**
+     * Column number starting from 1.
+     */
+    public int $column;
+}